requirements.txt
streamlit==1.40.1
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
agno==0.2.1
groq==0.9.0
yfinance==0.2.41
duckduckgo-search==3.9.10
plotly==5.24.0
pandas==2.2.0
numpy==1.24.3
ta-lib==0.4.28
technical-analysis==1.3.0
python-dateutil==2.8.2
scipy==1.13.0
scikit-learn==1.5.0